Separate out extensions methods
Should extensions now only contain extensions for Should
