+++ test/Qwiq.Tests.Common/Qwiq.Tests.Common.csproj
@@ -43,6 +43,7 @@
     </Compile>
     <Compile Include="Clock.cs" />
     <Compile Include="ContextSpecification.cs" />
+    <Compile Include="Extensions.cs" />
     <Compile Include="IContextSpecification.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ShouldExtensions.cs" />
